Renamed isMatch paramter to better reflect content and removed hardcoded file paths from main.
